dashboard_service/requirements.txt
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
requests